+++ models/consistency_group.go
@@ -61,7 +61,7 @@ type ConsistencyGroup struct {
 
 	// unique size
 	// Required: true
-	UniqueSize *float64 `json:"unique_size"`
+	UniqueSize *int64 `json:"unique_size"`
 }
 
 // Validate validates this consistency group
