+++ models/iscsi_lun_snapshot.go
@@ -54,7 +54,7 @@ type IscsiLunSnapshot struct {
 
 	// unique size
 	// Required: true
-	UniqueSize *float64 `json:"unique_size"`
+	UniqueSize *int64 `json:"unique_size"`
 }
 
 // Validate validates this iscsi lun snapshot
